Change logic for grabbing IP and displaying
The changes are in the core logic of the extension that allows it to dynamically build a list of interfaces with their IPv4 and IPv6 address.
The toggle will now not force update and will only cycle through the list of interfaces.
Switched getting WAN Public IP retrieval to `curl --max-time 5 -4 icanhazip.com` and `curl --max-time 5 -6 icanhazip.com` for IPv4 and IPv6 respectively because certain DNS configurations can prevent identifying the real Public IP address.
There are now extra checks to not display "None" when there is no IPv4 or IPv6 address or the Interface goes offline/disappears. This, also, means that it can now handle interfaces where there is only IPv4 or IPv6.